basic work on getting CSS classes reflected in the UI (JSP only)
- move getHighestCssClassLevel to MessagesUtils
- introduce basic CSS styles

fixes #2974
