+++ opengrok-web/src/main/webapp/default/style.css
@@ -912,19 +912,6 @@ a.scope            { /* scope              */ color: steelblue; font-weight: bol
     max-height: 480px;
 }
 
-.important-note {
-    background-color: rgb(255, 141, 96);
-    text-align: center;
-    display: inline-block;
-    max-height: 100%;
-}
-
-.important-note.important-note-rounded {
-    border-radius: 50%;
-    max-width: 100%;
-    min-width: 20px;
-}
-
 /**** DIFF NAVIGATION/JUMPER *****/
 
 .diff_navigation_style {
@@ -1247,6 +1234,34 @@ h6.message-group-caption {
 .message-group-item-text {
     margin-bottom: 0;
     line-height: 1.3;
+}
+
+.important-note {
+    text-align: center;
+    display: inline-block;
+    max-height: 100%;
+}
+
+.note-success {
+    background-color: #3c763d;
+}
+
+.note-info {
+    background-color: #31708f;
+}
+
+.note-warning {
+    background-color: #8a6d3b;
+}
+
+.note-error {
+    background-color: #a94442;
+}
+
+.important-note.important-note-rounded {
+    border-radius: 50%;
+    max-width: 100%;
+    min-width: 20px;
 }
 
 /** --------------- quick pull class ------------ */
