+++ opengrok-web/src/main/webapp/mast.jsp
@@ -98,7 +98,7 @@ include file="pageheader.jspf"
     }
     %>
     <% if (!messages.isEmpty()) { %>
-    <span class="important-note">
+    <span class="note-<%= MessagesUtils.getCssClass(cfg.getProject().getName(), MessagesContainer.MESSAGES_MAIN_PAGE_TAG) %> important-note">
     <% } %>
         <a href="<%= context + Prefix.XREF_P %>/">xref</a>: <%= Util
         .breadcrumbPath(context + Prefix.XREF_P, path,'/',"",true,cfg.isDir()) %>
@@ -114,7 +114,7 @@ include file="pageheader.jspf"
     %></span>
     <% if (!messages.isEmpty()) { %>
     </span>
-    <span class="important-note important-note-rounded"
+    <span class="note-<%= MessagesUtils.getCssClass(cfg.getProject().getName(), MessagesContainer.MESSAGES_MAIN_PAGE_TAG) %> important-note important-note-rounded"
           data-messages='<%= messages %>'>!</span>
     <% }
 }
